Nqsb_config/unix: Change how client certificate validation is configured
A client certificate must be asked if the context has a CA set and if certificate verification isn't disabled.
If no CA is set, then no certificate will be asked for.
If the context has been configured to not verify client/server certs (via tls_config_verify_client/server_cert), no certificate will be asked for.